update test setup
